feat: `viur install update` command (#103)
* fixed  merge conflict related bugs

* feature/updateAllForBuilds

* PEP8 Fixes

* refactored old install.py into packtet.py as a pseudo packet manager for viur tools

* Ai generated Documentation

* refactoring and  documentation

* Update README.md


---------
