+++ src/viur_cli/__init__.py
@@ -1,7 +1,7 @@
 from .cli import *
 from .conf import *
 from .deploy import *
-from .install import *
+from .package import *
 from .local import *
 from .build import *
 from .setup import *
@@ -9,3 +9,4 @@ from .version import *
 from .scriptor import script
 from .tool import *
 from .update import *
+from .package import *
